Feature/authorize

+++ src/main/java/team7/simple/domain/answer/service/AnswerService.java
@@ -6,10 +6,14 @@ import team7.simple.domain.answer.dto.AnswerRequestDto;
 import team7.simple.domain.answer.dto.AnswerResponseDto;
 import team7.simple.domain.answer.dto.AnswerUpdateParam;
 import team7.simple.domain.answer.entity.Answer;
+import team7.simple.domain.answer.error.exception.CAnswerNotFoundException;
 import team7.simple.domain.answer.repository.AnswerJpaRepository;
+import team7.simple.domain.auth.error.exception.CAccessDeniedException;
 import team7.simple.domain.question.entity.Question;
 import team7.simple.domain.question.service.QuestionService;
-import team7.simple.domain.answer.error.exception.CAnswerNotFoundException;
+import team7.simple.domain.user.entity.User;
+import team7.simple.domain.user.error.exception.CUserNotFoundException;
+import team7.simple.domain.user.repository.UserJpaRepository;
 
 import javax.transaction.Transactional;
 import java.util.List;
@@ -21,26 +25,34 @@ public class AnswerService {
     private final AnswerJpaRepository answerJpaRepository;
     private final QuestionService questionService;
 
+    private final UserJpaRepository userJpaRepository;
+
     @Transactional
-    public Long createAnswer(Long questionId, AnswerRequestDto answerRequestDto) {
+    public Long createAnswer(Long questionId, AnswerRequestDto answerRequestDto, User authUser) {
+        User user = userJpaRepository.findById(authUser.getId()).orElseThrow(CUserNotFoundException::new);
         Question question = questionService.getQuestionById(questionId);
-        Answer answer = answerRequestDto.toEntity(question);
+        Answer answer = answerRequestDto.toEntity(question, user);
 
         return answerJpaRepository.save(answer).getId();
     }
 
     @Transactional
-    public Long updateAnswer(AnswerUpdateParam answerUpdateParam) {
-        Long answerId = answerUpdateParam.getAnswerId();
+    public Long updateAnswer(Long answerId, AnswerUpdateParam answerUpdateParam, User user) {
         Answer answer = getAnswerById(answerId);
-        answer.setContent(answerUpdateParam.getContent());
+        if (!answer.getUser().getAccount().equals(user.getAccount())){
+            throw new CAccessDeniedException();
+        }
+        answer.update(answerUpdateParam.getContent());
 
         return answerId;
     }
 
     @Transactional
-    public void deleteAnswer(Long answerId) {
+    public void deleteAnswer(Long answerId, User user) {
         Answer answer = getAnswerById(answerId);
+        if (!answer.getUser().getAccount().equals(user.getAccount())){
+            throw new CAccessDeniedException();
+        }
         answerJpaRepository.delete(answer);
     }
