+++ src/main/java/team7/simple/domain/unit/service/UnitService.java
@@ -1,9 +1,11 @@
 package team7.simple.domain.unit.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+import team7.simple.domain.auth.error.exception.CAccessDeniedException;
 import team7.simple.domain.course.entity.Course;
 import team7.simple.domain.course.service.CourseService;
 import team7.simple.domain.file.entity.Video;
@@ -17,6 +19,7 @@ import team7.simple.domain.unit.repository.UnitJpaRepository;
 import team7.simple.domain.file.dto.VideoDto;
 import team7.simple.domain.file.service.VideoService;
 import team7.simple.domain.unit.error.exception.CUnitNotFoundException;
+import team7.simple.domain.user.entity.User;
 import team7.simple.infra.hls.service.HlsService;
 
 import java.util.List;
@@ -24,6 +27,7 @@ import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
+@Slf4j
 public class UnitService {
     private final UnitJpaRepository unitJpaRepository;
     private final CourseService courseService;
@@ -54,23 +58,27 @@ public class UnitService {
     }
 
     @Transactional
-    public void deleteUnit(Long unitId) {
+    public void deleteUnit(Long unitId, User user) {
         Unit unit = unitJpaRepository.findById(unitId).orElseThrow(CUnitNotFoundException::new);
+        if (!unit.getCourse().getInstructor().getAccount().equals(user.getAccount())) {
+            throw new CAccessDeniedException();
+        }
         unitJpaRepository.delete(unit);
     }
 
     @Transactional
-    public Long updateUnit(Long unitId, UnitUpdateParam unitUpdateParam) {
+    public Long updateUnit(Long unitId, UnitUpdateParam unitUpdateParam, User user) {
         Unit unit = getUnitById(unitId);
-        unit.setTitle(unitUpdateParam.getTitle());
-        unit.setDescription(unitUpdateParam.getDescription());
-        unit.setObjective(unit.getObjective());
+        if (!unit.getCourse().getInstructor().getAccount().equals(user.getAccount())) {
+            throw new CAccessDeniedException();
+        }
+        unit.update(unitUpdateParam.getTitle(), unitUpdateParam.getDescription(), unitUpdateParam.getObjective());
 
         return unitId;
     }
 
     @Transactional
-    public Unit getUnitById(Long unitId){
+    public Unit getUnitById(Long unitId) {
         return unitJpaRepository.findById(unitId).orElseThrow(CUnitNotFoundException::new);
     }
 
@@ -84,6 +92,7 @@ public class UnitService {
 
     /**
      * 강의의 세부 정보를 반환합니다.
+     *
      * @param unitId 유닛아이디
      * @return UnitDetailResponseDto (유닛아이디, 제목, 강의 소개, 강의 목표)
      */
